Prevent spam-click of return home button
